Updates for the Mental Health (Discrimination) Act 2013
"Amendments to model articles

The model articles were amended by the Mental Health (Discrimination) Act 2013 on 28 April 2013 to remove the provision for terminating a director’s appointment on grounds of mental health. This provision could be found in:

    paragraph 18(e) of the model articles for private companies limited by shares or by guarantee
    paragraph 22(e) of the model articles for public companies

You don’t have to remove the provision if it’s in your articles, but you can do so by amending your articles or by adopting the newer model articles."

In section 3, Company directors, omit Schedule 2, paragraph 18(e).