Director term
A proposal that where there are three or fewer members standing as a director at an AGM, a director at the end of a three-year term may be reelected.
At present, we have 4 directors who if reelected will be in the final year of their board membership. In previous years, no new candidates have come forward as directors. This has a serious risk that we fall below quorum of 3 and the hacklab is crippled, even if an existing director was willing to stay on. In 2020, it took 6 months for members to come forward and volunteer which could be a serious issue for the lab