Election notification option B
As per River's proposal this version requires notification of intent to stand. However, in the event that there is not an election, clause 70 remains. This is important as all directors must stand down at an AGM and be reelected so if nobody came forward specifically within the proscribed period we could end up with no directors.

Clause 70 is 

"A director retiring at an AGM will be deemed to have been re-elected unless:
70.1 they advise the board prior to the conclusion of the AGM that they do not wish to be reappointed as a director; or
70.2 an election process was held at the AGM and they were not among those elected/re-elected through that process."